requirements.txt
# Telegram Bot
python-telegram-bot==21.3

# AI Providers
openai==1.40.0
anthropic==0.34.0

# For local models (optional)
# ollama==0.3.0

# Environment and utilities
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.4.0

# Database
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.30
greenlet==3.0.3

# Scheduler for background tasks
apscheduler==3.10.4

# Date/time utilities
python-dateutil==2.9.0

# HTTP client for TON API
aiohttp==3.9.5

# QR code generation
qrcode==7.4.2
Pillow==10.3.0
